CXFSA-540-1: Removed all store implementation for premium calendar
