+++ projects/fsastorefrontlib/src/cms-components/my-account/premium-calendar/premium-calendar.component.spec.ts
@@ -1,11 +1,20 @@
 import { waitForAsync, ComponentFixture, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { I18nTestingModule, OccConfig } from '@spartacus/core';
+import {
+  I18nTestingModule,
+  UserIdService,
+  OCC_USER_ID_CURRENT,
+} from '@spartacus/core';
 import { SpinnerModule } from '@spartacus/storefront';
 import { PolicyService } from '../../../core/my-account/facade';
+import { PolicyConnector } from '../../../core/my-account/connectors/policy.connector';
 import { PremiumCalendarComponent } from './premium-calendar.component';
 import createSpy = jasmine.createSpy;
-import { of } from 'rxjs';
+import { Observable, of } from 'rxjs';
+
+const policyId = 'PL00001';
+const contractId = 'CT00001';
+const testCategory = 'testCategory';
 
 const policy1 = {
   policyId: 'policyId',
@@ -21,6 +30,23 @@ const policies = {
   insurancePolicies: [policy1],
 };
 
+const mockedPolicy = {
+  policyId: policyId,
+  contractId: contractId,
+};
+
+class MockPolicyConnector {
+  getPremiumCalendar() {
+    return of(mockedPolicy);
+  }
+}
+
+class MockUserIdService {
+  getUserId(): Observable<string> {
+    return of(OCC_USER_ID_CURRENT);
+  }
+}
+
 class MockPolicyService {
   loadPremiumCalendar = createSpy();
   getPremiumCalendarLoaded = createSpy();
@@ -33,6 +59,8 @@ describe('PremiumCalendarComponent', () => {
   let component: PremiumCalendarComponent;
   let fixture: ComponentFixture<PremiumCalendarComponent>;
   let policyService: PolicyService;
+  let userIdService: MockUserIdService;
+  let policyConnector: PolicyConnector;
 
   beforeEach(
     waitForAsync(() => {
@@ -44,16 +72,23 @@ describe('PremiumCalendarComponent', () => {
             provide: PolicyService,
             useClass: MockPolicyService,
           },
+          { provide: UserIdService, useClass: MockUserIdService },
         ],
       }).compileComponents();
     })
   );
 
   beforeEach(() => {
+    policyService = TestBed.inject(PolicyService);
+    policyConnector = TestBed.inject(PolicyConnector);
+    userIdService = TestBed.inject(UserIdService);
     fixture = TestBed.createComponent(PremiumCalendarComponent);
     component = fixture.componentInstance;
+
+    spyOn(policyConnector, 'getPremiumCalendar').and.callThrough();
+    spyOn(userIdService, 'getUserId').and.callThrough();
+
     fixture.detectChanges();
-    policyService = TestBed.inject(PolicyService);
   });
 
   it('should create', () => {
