+++ projects/fsastorefrontlib/src/cms-components/my-account/premium-calendar/premium-calendar.component.ts
@@ -1,6 +1,9 @@
 import { ChangeDetectionStrategy, Component, OnDestroy } from '@angular/core';
 import { PolicyService } from '../../../core/my-account/facade';
 import { Subscription } from 'rxjs';
+import { PolicyConnector } from '../../../core/my-account/connectors/policy.connector';
+import { UserIdService } from '@spartacus/core';
+import { switchMap } from 'rxjs/operators';
 
 @Component({
   selector: 'cx-fs-premium-calendar',
@@ -8,9 +11,16 @@ import { Subscription } from 'rxjs';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class PremiumCalendarComponent implements OnDestroy {
-  constructor(protected policyService: PolicyService) {}
+  constructor(
+    protected policyService: PolicyService,
+    protected userIdService: UserIdService,
+    protected policyConnector: PolicyConnector
+  ) {}
 
-  policies$ = this.policyService.getPremiumCalendar();
+  policies$ = this.userIdService
+    .getUserId()
+    .pipe(switchMap(userId => this.policyConnector.getPremiumCalendar(userId)));
+
   selectedIndexes: number[] = [];
   private subscription = new Subscription();
